Mid-term version: report, modified testcases, better error/throw

+++ build.sh
@@ -1,7 +1,9 @@
 #!/bin/bash
 
 cd src
-rm -r build
+if [ -d "build" ]; then
+    rm -r build
+fi
 mkdir -p build && cd build
 cmake ..
 make -j8
